id type changed (int -> Long)

+++ src/main/java/ua/com/foxminded/university/dao/CourseDao.java
@@ -3,6 +3,6 @@ package ua.com.foxminded.university.dao;
 import ua.com.foxminded.university.model.Course;
 
 public interface CourseDao extends GenericDao<Course> {
-    public Course getByIdWithDetail(int id);
+    public Course getByIdWithDetail(Long id);
     public int updateGroups(Course item);
 }
