+++ src/main/java/ua/com/foxminded/university/dao/LessonDao.java
@@ -7,9 +7,9 @@ import java.util.Optional;
 import ua.com.foxminded.university.model.Lesson;
 
 public interface LessonDao extends GenericDao<Lesson> {
-    public List<Lesson> getByCourseId(int curseId);
-    public List<Lesson> getByGroupId(int groupId);
-    public List<Lesson> getByTeacherId(int teacherId);
-    public Optional<Lesson> getByDatePeriodIdTeacherId(LocalDate date, int periodId, int teacherId);
-    public Optional<Lesson> getByDatePeriodIdClassroomId(LocalDate date, int periodId, int classroomId);
+    public List<Lesson> getByCourseId(Long curseId);
+    public List<Lesson> getByGroupId(Long groupId);
+    public List<Lesson> getByTeacherId(Long teacherId);
+    public Optional<Lesson> getByDatePeriodIdTeacherId(LocalDate date, Long periodId, Long teacherId);
+    public Optional<Lesson> getByDatePeriodIdClassroomId(LocalDate date, Long periodId, Long classroomId);
 }
