+++ src/main/java/ua/com/foxminded/university/service/LessonService.java
@@ -5,7 +5,7 @@ import java.util.List;
 import ua.com.foxminded.university.model.Lesson;
 
 public interface LessonService extends GenericService<Lesson>{
-    public List<Lesson> getByCourseId(int curseId);
-    public List<Lesson> getByGroupId(int groupId);
-    public List<Lesson> getByTeacherId(int teacherId);
+    public List<Lesson> getByCourseId(Long curseId);
+    public List<Lesson> getByGroupId(Long groupId);
+    public List<Lesson> getByTeacherId(Long teacherId);
 }
